# docs/openapi.yaml
openapi: 3.0.3
info:
  title: パートナー協業プラットフォーム API
  description: |
    パートナー企業との協業を管理するためのREST API仕様です。

    ## 認証
    このAPIはBearer Token (JWT)認証を使用します。
    ログインAPIで取得したアクセストークンを`Authorization`ヘッダーに設定してください。

    ```
    Authorization: Bearer {accessToken}
    ```

    ## ページネーション
    一覧取得APIでは`limit`と`offset`パラメータでページネーションを制御できます。

    ## エラーレスポンス
    全てのエラーは統一フォーマットで返却されます。
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  - url: https://api.example.com/api
    description: Production server
  - url: https://staging-api.example.com/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /partners:
    get:
      tags:
        - Partners
      summary: パートナー一覧取得
      description: パートナー一覧を取得します。
      operationId: listPartners
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          description: ステータスでフィルタ
          schema:
            $ref: '#/components/schemas/PartnerStatus'
        - name: search
          in: query
          description: 会社名・担当者名での検索
          schema:
            type: string
        - name: industry
          in: query
          description: 業種でフィルタ
          schema:
            type: string
        - name: rating
          in: query
          description: 評価（この値以上）でフィルタ
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: sortBy
          in: query
          description: ソート項目
          schema:
            type: string
            enum: [createdAt, name, rating]
            default: createdAt
        - name: sortOrder
          in: query
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Partners
      summary: パートナー登録
      description: 新規パートナーを登録します。
      operationId: createPartner
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerCreateRequest'
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Partner'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /partners/{id}:
    get:
      tags:
        - Partners
      summary: パートナー詳細取得
      description: 指定したパートナーの詳細情報を取得します。
      operationId: getPartner
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PartnerIdParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PartnerDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Partners
      summary: パートナー更新
      description: 既存のパートナー情報を更新します。
      operationId: updatePartner
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PartnerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Partner'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ============================================================
  # Dashboard APIs
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTアクセストークン
  parameters:
    LimitParam:
      name: limit
      in: query
      description: 取得件数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: 取得開始位置
      schema:
        type: integer
        minimum: 0
        default: 0
...
    PartnerIdParam:
      name: id
      in: path
      required: true
      description: パートナーID
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: 認証が必要です
...
    NotFound:
      description: リソース未検出
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: 指定されたリソースが見つかりません
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: 入力内容に誤りがあります
              details:
                - field: email
                  message: 有効なメールアドレスを入力してください
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: フィールド名
                  message:
                    type: string
                    description: 詳細メッセージ
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
        limit:
          type: integer
          description: 取得件数
        offset:
          type: integer
          description: 取得開始位置
        hasMore:
          type: boolean
          description: 追加データの有無
...
    ProjectStatus:
      type: string
      enum: [draft, active, completed, archived]
      description: 案件ステータス
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
          description: 案件ID
        name:
          type: string
          description: 案件名
        status:
          $ref: '#/components/schemas/ProjectStatus'
...
    PartnerStatus:
      type: string
      enum: [pending, active, inactive]
      description: パートナーステータス
...
    Partner:
      type: object
      properties:
        id:
          type: string
          description: パートナーID
        name:
          type: string
          description: 会社名
        industry:
          type: string
          description: 業種
        status:
          $ref: '#/components/schemas/PartnerStatus'
        contactPerson:
          type: string
          description: 担当者名
        email:
          type: string
          format: email
          description: メールアドレス
        phone:
          type: string
          description: 電話番号
        rating:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: 評価
        totalProjects:
          type: integer
          description: 総案件数
        activeProjects:
          type: integer
          description: 進行中案件数
        contractStartDate:
          type: string
          format: date
          description: 契約開始日
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
    PartnerDetail:
      allOf:
        - $ref: '#/components/schemas/Partner'
        - type: object
          properties:
            description:
              type: string
              description: 会社説明
            address:
              $ref: '#/components/schemas/Address'
            website:
              type: string
              format: uri
              description: WebサイトURL
            skills:
              type: array
              items:
                type: string
              description: スキル・技術
            certifications:
              type: array
              items:
                type: string
              description: 資格・認定
            contractInfo:
              $ref: '#/components/schemas/ContractInfo'
            statistics:
              $ref: '#/components/schemas/PartnerStatistics'
            recentProjects:
              type: array
              items:
                $ref: '#/components/schemas/ProjectSummary'
              description: 最近の案件
            notes:
              type: string
              description: 備考
    PartnerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Partner'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PartnerCreateRequest:
      type: object
      required:
        - name
        - industry
        - contactPerson
        - email
      properties:
        name:
          type: string
          maxLength: 100
          description: 会社名
        industry:
          type: string
          description: 業種
        status:
          type: string
          enum: [pending, active]
          default: pending
          description: ステータス
        description:
          type: string
          maxLength: 2000
          description: 会社説明
        address:
          $ref: '#/components/schemas/Address'
        contactPerson:
          type: string
          description: 担当者名
        email:
          type: string
          format: email
          description: メールアドレス
        phone:
          type: string
          description: 電話番号
        website:
          type: string
          format: uri
          description: WebサイトURL
        skills:
          type: array
          items:
            type: string
          description: スキル・技術
        certifications:
          type: array
          items:
            type: string
          description: 資格・認定
        contractInfo:
          $ref: '#/components/schemas/ContractInfo'
        notes:
          type: string
          description: 備考
    PartnerUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: 会社名
        industry:
          type: string
          description: 業種
        status:
          $ref: '#/components/schemas/PartnerStatus'
        description:
          type: string
          maxLength: 2000
          description: 会社説明
        address:
          $ref: '#/components/schemas/Address'
        contactPerson:
          type: string
          description: 担当者名
        email:
          type: string
          format: email
          description: メールアドレス
        phone:
          type: string
          description: 電話番号
        website:
          type: string
          format: uri
          description: WebサイトURL
        skills:
          type: array
          items:
            type: string
          description: スキル・技術
        certifications:
          type: array
          items:
            type: string
          description: 資格・認定
        contractInfo:
          $ref: '#/components/schemas/ContractInfo'
        rating:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          description: 評価
        notes:
          type: string
          description: 備考
    Address:
      type: object
      properties:
        postalCode:
          type: string
          description: 郵便番号
        prefecture:
          type: string
          description: 都道府県
        city:
          type: string
          description: 市区町村
        street:
          type: string
          description: 町名・番地
        building:
          type: string
          description: 建物名
    ContractInfo:
      type: object
      properties:
        contractStartDate:
          type: string
          format: date
          description: 契約開始日
        contractEndDate:
          type: string
          format: date
          description: 契約終了日
        contractType:
          type: string
          description: 契約タイプ
        paymentTerms:
          type: string
          description: 支払条件
    PartnerStatistics:
      type: object
      properties:
        totalProjects:
          type: integer
          description: 総案件数
        activeProjects:
          type: integer
          description: 進行中案件数
        completedProjects:
          type: integer
          description: 完了案件数
        averageProjectDuration:
          type: integer
          description: 平均案件期間（日）
        onTimeDeliveryRate:
          type: number
          format: float
          description: 納期遵守率（%）

    # ============================================================
    # Dashboard Schemas
    # ============================================================
...